refactor(ThreadsAlgo): Move Threads algorithm to its own RefreshController instead of being in MessageController
