+++ app/src/main/java/com/infomaniak/mail/workers/DraftsActionsWorker.kt
@@ -33,8 +33,8 @@ import com.infomaniak.mail.data.api.ApiRoutes
 import com.infomaniak.mail.data.cache.RealmDatabase
 import com.infomaniak.mail.data.cache.mailboxContent.DraftController
 import com.infomaniak.mail.data.cache.mailboxContent.FolderController
-import com.infomaniak.mail.data.cache.mailboxContent.MessageController
-import com.infomaniak.mail.data.cache.mailboxContent.MessageController.RefreshMode
+import com.infomaniak.mail.data.cache.mailboxContent.RefreshController
+import com.infomaniak.mail.data.cache.mailboxContent.RefreshController.RefreshMode
 import com.infomaniak.mail.data.cache.mailboxInfo.MailboxController
 import com.infomaniak.mail.data.models.AppSettings
 import com.infomaniak.mail.data.models.Attachment
@@ -203,7 +203,7 @@ class DraftsActionsWorker @AssistedInject constructor(
             if (times.isNotEmpty()) delay += max(times.maxOf { it } - timeNow, 0L)
             delay(min(delay, MAX_REFRESH_DELAY))
 
-            MessageController.refreshThreads(
+            RefreshController.refreshThreads(
                 refreshMode = RefreshMode.NEW_MESSAGES_WITH_ROLE,
                 mailbox = mailbox,
                 folder = folder,
